[VS Incentives] priviledged CreateGroupAsIncentivesModuleAcc for upgrade handler
