+++ x/incentives/types/errors.go
@@ -9,6 +9,7 @@ import (
 var (
 	ErrNoPoolIDsGiven        = fmt.Errorf("no pool IDs given")
 	ErrZeroNumEpochsPaidOver = fmt.Errorf("num epochs paid over must be greater than zero for non-perpetual gauges")
+	ErrUnauthorized          = fmt.Errorf("unauthorized to perform this action. Must be an incentives module account")
 )
 
 type UnsupportedSplittingPolicyError struct {
